# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Parbati Enterprises Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh JWT token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: New token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/AuthUser"
                required: [user]
...
  /expenses/{id}/documents:
    get:
      operationId: listExpenseDocuments
      tags: [documents]
      summary: List documents for the expense's current version
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
                required: [data]
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: uploadExpenseDocuments
      tags: [documents]
      summary: Upload documents to the expense's current version (expense_entry+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UploadedFile"
                required: [data]
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{expenseId}/versions/{versionId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for an expense version
      security:
        - bearerAuth: []
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
                required: [data]
    post:
      operationId: uploadDocuments
      tags: [documents]
      summary: Upload documents for an expense version (expense_entry+)
      security:
        - bearerAuth: []
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UploadedFile"
                required: [data]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [expense_entry, accounts, admin, superadmin]
        canViewHistory:
          type: boolean
      required: [id, email, name, role, canViewHistory]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [token, user]
...
    Document:
      type: object
      properties:
        id:
          type: integer
        expenseVersionId:
          type: integer
        voucherNumber:
          type: string
        originalName:
          type: string
        storedName:
          type: string
        mimeType:
          type: string
          nullable: true
        fileSize:
          type: integer
        publicUrl:
          type: string
          nullable: true
        fileOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - expenseVersionId
        - voucherNumber
        - originalName
        - storedName
        - fileSize
        - fileOrder
    UploadedFile:
      type: object
      properties:
        id:
          type: integer
        originalName:
          type: string
        publicUrl:
          type: string
      required: [id, originalName, publicUrl]
...